Colocate auth schemas and extract primitives to lib/schemas
The auth schemas (emailSchema, passwordSchema, loginSchema, tokenSchema,
resetPasswordSchema, changeEmailSchema, changePasswordSchema,
requestResetSchema) lived in `lib/utils/schema.ts` and were imported
by 4 different page.server.ts files. They were too far from their
consumers and hard to discover in context.

After this refactor:
- `emailSchema` and `passwordSchema` move to
  `web-v4/src/lib/schemas/auth.ts` as shared primitives. They are
  genuinely cross-page (used by 3+ features) and worth a small
  shared module.
- The page-specific schemas (login, token, requestReset, resetPassword,
  changeEmail, changePassword) move into the page that uses them,
  following the colocation principle.
- `lib/utils/schema.ts` shrinks to just the `parseForm` helper.

Subsequent schema migrations for the vehicles, reminders, and tasks
actions will colocate their schemas in the same way.
