+++ web-v4/src/routes/reset-password/[token]/+page.server.ts
@@ -2,9 +2,25 @@ import { resetPassword } from '$lib/data/session';
 import { setSession } from '$lib/utils/session';
 import { HTTPError } from '$lib/data/client';
 import { withActionErrors } from '$lib/utils/actions';
-import { parseForm, resetPasswordSchema } from '$lib/utils/schema';
+import { parseForm } from '$lib/utils/schema';
 import { fail, redirect } from '@sveltejs/kit';
+import { passwordSchema } from '$lib/schemas/auth';
+import * as v from 'valibot';
 import type { Actions, PageServerLoad } from './$types';
+
+const passwordMatchCheck = v.check(
+	(value: { password: string; confirm_password: string }) =>
+		value.password === value.confirm_password,
+	'Passwords do not match',
+);
+
+const resetPasswordSchema = v.pipe(
+	v.object({
+		password: passwordSchema,
+		confirm_password: v.string('Password confirmation is required'),
+	}),
+	v.forward(passwordMatchCheck, ['confirm_password']),
+);
 
 export const load: PageServerLoad = async ({ params, locals }) => {
 	// If already logged in, go to vehicles
